first round: move jit_row to submodule, separate validity and layout
