+++ datafusion/core/src/row/layout.rs
@@ -1,0 +1,67 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+//! Various row layout for different use case
+
+use crate::row::{schema_null_free, var_length};
+use arrow::datatypes::{DataType, Schema};
+use arrow::util::bit_util::{ceil, round_upto_power_of_2};
+use std::sync::Arc;
+
+const UTF8_DEFAULT_SIZE: usize = 20;
+const BINARY_DEFAULT_SIZE: usize = 100;
+
+/// Get relative offsets for each field and total width for values
+pub fn get_offsets(null_width: usize, schema: &Arc<Schema>) -> (Vec<usize>, usize) {
+    let mut offsets = vec![];
+    let mut offset = null_width;
+    for f in schema.fields() {
+        offsets.push(offset);
+        offset += type_width(f.data_type());
+    }
+    (offsets, offset - null_width)
+}
+
+fn type_width(dt: &DataType) -> usize {
+    use DataType::*;
+    if var_length(dt) {
+        return std::mem::size_of::<u64>();
+    }
+    match dt {
+        Boolean | UInt8 | Int8 => 1,
+        UInt16 | Int16 => 2,
+        UInt32 | Int32 | Float32 | Date32 => 4,
+        UInt64 | Int64 | Float64 | Date64 => 8,
+        _ => unreachable!(),
+    }
+}
+
+/// Estimate row width based on schema
+pub fn estimate_row_width(schema: &Arc<Schema>) -> usize {
+    let null_free = schema_null_free(schema);
+    let field_count = schema.fields().len();
+    let mut width = if null_free { 0 } else { ceil(field_count, 8) };
+    for f in schema.fields() {
+        width += type_width(f.data_type());
+        match f.data_type() {
+            DataType::Utf8 => width += UTF8_DEFAULT_SIZE,
+            DataType::Binary => width += BINARY_DEFAULT_SIZE,
+            _ => {}
+        }
+    }
+    round_upto_power_of_2(width, 8)
+}
